+++ mycodo/config_devices_units.py
@@ -61,6 +61,10 @@ MEASUREMENTS = {
         'name': lazy_gettext('Disk'),
         'meas': 'disk_space',
         'units': ['MB', 'kB', 'GB']},
+    'dissolved_oxygen': {
+        'name': lazy_gettext('Dissolved Oxygen'),
+        'meas': 'dissolved_oxygen',
+        'units': ['mg_L']},
     'duration_time': {
         'name': lazy_gettext('Duration'),
         'meas': 'duration_time',
@@ -84,7 +88,7 @@ MEASUREMENTS = {
     'electrical_potential': {
         'name': lazy_gettext('Electrical Potential'),
         'meas': 'electrical_potential',
-        'units': ['V']},
+        'units': ['V', 'mV']},
     'frequency': {
         'name': lazy_gettext('Frequency'),
         'meas': 'frequency',
@@ -113,6 +117,10 @@ MEASUREMENTS = {
         'name': lazy_gettext('Moisture'),
         'meas': 'moisture',
         'units': ['unitless']},
+    'oxidation_reduction_potential': {
+        'name': lazy_gettext('Oxidation Reduction Potential'),
+        'meas': 'oxidation_reduction_potential',
+        'units': ['mV', 'V']},
     'particulate_matter_1_0': {
         'name': 'PM1',
         'meas': 'particulate_matter_1_0',
@@ -258,9 +266,15 @@ UNITS = {
     'm': {
         'name': lazy_gettext('Meter'),
         'unit': 'm'},
+    'mg_L': {
+        'name': lazy_gettext('Milligram per Liter'),
+        'unit': 'mg/L'},
     'ml': {
         'name': lazy_gettext('Milliliter'),
         'unit': 'ml'},
+    'mV': {
+        'name': lazy_gettext('Millivolt'),
+        'unit': 'mV'},
     'm3_kg': {
         'name': lazy_gettext('Cubic meters per kilogram'),
         'unit': 'm^3/kg'},
@@ -358,5 +372,9 @@ UNIT_CONVERSIONS = [
 
     # Time
     ('s', 'minute', 'x/60'),
-    ('minute', 's', 'x*60')
+    ('minute', 's', 'x*60'),
+
+    # Volt
+    ('V', 'mV', 'x*1000'),
+    ('mV', 'V', 'x/1000')
 ]